updated mipmapping mechanism; removed black grid from Windows version
git-svn-id: https://paroxysm.googlecode.com/svn/trunk@320 1a4693bc-ef48-11de-b206-85b67c77eac7

+++ source/TerrainGrid.h
@@ -63,7 +63,7 @@ class TerrainGrid
         Matrix<float> mHeights;
         GLfloat* mVertices;
         GLfloat* mNormals;
-        GLfloat* mTextureCoordinates;
+        GLint* mTextureCoordinates;
         GLuint* mIndices;
         GLsizei mNumIndices;
         GLuint mNumVerticesX3;
